+++ CHANGELOG.md
@@ -10,6 +10,42 @@
 
 ---
 
+# v1.1.3 (Fri Mar 21 2025)
+
+#### ⚠️ Pushed to `main`
+
+- fix: refresh correctly tokens ([@Sebas3245](https://github.com/Sebas3245))
+
+#### Authors: 1
+
+- Juan Sebastian Lara Aros ([@Sebas3245](https://github.com/Sebas3245))
+
+---
+
+# v1.1.2 (Thu Mar 20 2025)
+
+#### ⚠️ Pushed to `main`
+
+- feat: added reset timeout by change page ([@Sebas3245](https://github.com/Sebas3245))
+
+#### Authors: 1
+
+- Juan Sebastian Lara Aros ([@Sebas3245](https://github.com/Sebas3245))
+
+---
+
+# v1.1.1 (Wed Mar 19 2025)
+
+#### ⚠️ Pushed to `main`
+
+- feat: added critial paths ([@Sebas3245](https://github.com/Sebas3245))
+
+#### Authors: 1
+
+- Juan Sebastian Lara Aros ([@Sebas3245](https://github.com/Sebas3245))
+
+---
+
 # v1.1.0 (Wed Mar 19 2025)
 
 #### 💥 Breaking Change